pyproject.toml
[tool.poetry]
name = "website-image-scraper"
version = "0.1.0"
description = "A web scraper to download images from a website"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9,<3.13"
requests = "^2.31.0"
beautifulsoup4 = "^4.12.3"
rembg = "^2.0.60"
Pillow = "^10.1.0"
onnxruntime = "^1.16.1"
opencv-python-headless = "^4.8.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
flake8 = "^7.0.0"
black = "^24.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"